refactor HubConfig to separate out general hub functionality from viz-specific functionality. precursor to [add support for multiple model tasks #38]
