+++ tests/hub_predtimechart/test_generate_data.py
@@ -5,7 +5,7 @@ import pandas as pd
 
 from hub_predtimechart.app.generate_json_files import json_file_name
 from hub_predtimechart.generate_data import forecast_data_for_model_df
-from hub_predtimechart.hub_config import HubConfig
+from hub_predtimechart.hub_config_ptc import HubConfigPtc
 
 
 def test_json_file_name():
@@ -18,7 +18,7 @@ def test_json_file_name():
 
 def test_forecast_data_for_model_df_complex_forecast_hub_US():
     hub_dir = Path('tests/hubs/example-complex-forecast-hub')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
 
     with open('tests/expected/example-complex-forecast-hub/forecasts/wk-inc-flu-hosp_US_2022-10-22.json') as fp:
         exp_data = json.load(fp)
@@ -43,7 +43,7 @@ def test_forecast_data_for_model_df_complex_forecast_hub_US():
 
 def test_forecast_data_for_model_df_complex_forecast_hub_01():
     hub_dir = Path('tests/hubs/example-complex-forecast-hub')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
 
     with open('tests/expected/example-complex-forecast-hub/forecasts/wk-inc-flu-hosp_01_2022-10-22.json') as fp:
         exp_data = json.load(fp)
@@ -68,7 +68,7 @@ def test_forecast_data_for_model_df_complex_forecast_hub_01():
 
 def test_forecast_data_for_model_df_no_data():
     hub_dir = Path('tests/hubs/example-complex-forecast-hub')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
 
     model_output_file = hub_dir / 'model-output/Flusight-baseline/2022-10-22-Flusight-baseline.csv'
     act_data = forecast_data_for_model_df(hub_config, pd.read_csv(model_output_file), 'wk inc flu hosp', ('01',))
