+++ tests/hub_predtimechart/test_generate_json_files.py
@@ -3,7 +3,7 @@ import shutil
 from pathlib import Path
 
 from hub_predtimechart.app.generate_json_files import _generate_json_files, _generate_options_file
-from hub_predtimechart.hub_config import HubConfig
+from hub_predtimechart.hub_config_ptc import HubConfigPtc
 
 
 def test_generate_json_files(tmp_path):
@@ -11,7 +11,7 @@ def test_generate_json_files(tmp_path):
     An integration test of `generate_json_files.py`'s `_generate_json_files()`.
     """
     hub_dir = Path('tests/hubs/example-complex-forecast-hub')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
     output_dir = tmp_path
     json_files = _generate_json_files(hub_config, output_dir)
     assert set(json_files) == {output_dir / 'wk-inc-flu-hosp_US_2022-10-22.json',
@@ -34,7 +34,7 @@ def test_generate_json_files_skip_files(tmp_path):
     This validates that only new data will be generated
     """
     hub_dir = Path('tests/hubs/example-complex-forecast-hub')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
     output_dir = tmp_path
 
     # copy all but one prediction to the output directory
@@ -67,7 +67,7 @@ def test_generate_json_files_regenerate(tmp_path):
     This validates that only new data will be generated
     """
     hub_dir = Path('tests/hubs/example-complex-forecast-hub')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
     output_dir = tmp_path
 
     # copy all but one prediction to the output directory
@@ -88,7 +88,7 @@ def test_generate_options_file(tmp_path):
     An integration test of `generate_json_files.py`'s `_generate_options_file()`.
     """
     hub_dir = Path('tests/hubs/example-complex-forecast-hub')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
     ptc_options = tmp_path / 'ptc_options'
     _generate_options_file(hub_config, ptc_options)
     with open(ptc_options) as act_options_fp, \
