+++ tests/hub_predtimechart/test_generate_options.py
@@ -2,12 +2,12 @@ import json
 from pathlib import Path
 
 from hub_predtimechart.generate_options import ptc_options_for_hub
-from hub_predtimechart.hub_config import HubConfig
+from hub_predtimechart.hub_config_ptc import HubConfigPtc
 
 
 def test_generate_options_complex_forecast_hub():
     hub_dir = Path('tests/hubs/example-complex-forecast-hub')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
     with open('tests/expected/example-complex-forecast-hub/predtimechart-options.json') as fp:
         exp_options = json.load(fp)
     act_options = ptc_options_for_hub(hub_config)
@@ -22,7 +22,7 @@ def test_generate_options_complex_forecast_hub():
 
 def test_generate_options_complex_forecast_hub_no_disclaimer():
     hub_dir = Path('tests/hubs/example-complex-forecast-hub')
-    hub_config = HubConfig(hub_dir, Path('tests/configs/example-complex-no-disclaimer.yml'))
+    hub_config = HubConfigPtc(hub_dir, Path('tests/configs/example-complex-no-disclaimer.yml'))
     with open('tests/expected/example-complex-forecast-hub/predtimechart-options.json') as fp:
         exp_options = json.load(fp)
     act_options = ptc_options_for_hub(hub_config)
@@ -31,7 +31,7 @@ def test_generate_options_complex_forecast_hub_no_disclaimer():
 
 def test_generate_options_flusight_forecast_hub():
     hub_dir = Path('tests/hubs/FluSight-forecast-hub')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
     with open('tests/expected/FluSight-forecast-hub/predtimechart-options.json') as fp:
         exp_options = json.load(fp)
     act_options = ptc_options_for_hub(hub_config)
@@ -40,7 +40,7 @@ def test_generate_options_flusight_forecast_hub():
 
 def test_generate_options_flu_metrocast():
     hub_dir = Path('tests/hubs/flu-metrocast')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
     with open('tests/expected/flu-metrocast/predtimechart-options.json') as fp:
         exp_options = json.load(fp)
     act_options = ptc_options_for_hub(hub_config)
@@ -49,7 +49,7 @@ def test_generate_options_flu_metrocast():
 
 def test_generate_options_task_id_text_covid19_forecast_hub():
     hub_dir = Path('tests/hubs/covid19-forecast-hub')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
     with open('tests/expected/covid19-forecast-hub/predtimechart-options.json') as fp:
         exp_options = json.load(fp)
     act_options = ptc_options_for_hub(hub_config)
