+++ tests/hub_predtimechart/test_generate_target_data.py
@@ -8,12 +8,12 @@ from freezegun import freeze_time
 
 from hub_predtimechart.app.generate_target_json_files import reference_date_from_today, ptc_target_data, \
     _max_as_of_le__reference_date
-from hub_predtimechart.hub_config import HubConfig
+from hub_predtimechart.hub_config_ptc import HubConfigPtc
 
 
 def test_ptc_target_data_flusight_forecast_hub():
     hub_dir = Path('tests/hubs/FluSight-forecast-hub')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
     target_data_df = hub_config.get_target_data_df()
     for loc in ['US', '01']:
         task_ids_tuple = (loc,)
@@ -25,7 +25,7 @@ def test_ptc_target_data_flusight_forecast_hub():
 
 def test_ptc_target_data_covid19_forecast_hub():
     hub_dir = Path('tests/hubs/covid19-forecast-hub')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
     target_data_df = hub_config.get_target_data_df()
     for loc in ['US', '01']:
         task_ids_tuple = (loc,)
@@ -37,7 +37,7 @@ def test_ptc_target_data_covid19_forecast_hub():
 
 def test_ptc_target_data_flu_metrocast():
     hub_dir = Path('tests/hubs/flu-metrocast')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
     target_data_df = hub_config.get_target_data_df()
     # 2025-02-25 is the newest as_of in tests/hubs/flu-metrocast time-series.csv. we use 2025-03-01 for
     # reference_date b/c that's the one right after 2025-02-25 (but we could have used any reference_date after
@@ -69,7 +69,7 @@ def test_reference_date_from_today():
 
 def test_get_target_data_df_error_cases():
     hub_dir = Path('tests/hubs/FluSight-forecast-hub')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
     hub_config.target_data_file_name = 'target-hospital-admissions-no-na.csv'  # override
     act_target_data_df = hub_config.get_target_data_df()
     assert act_target_data_df['value'].dtype == pl.datatypes.Float64
@@ -88,7 +88,7 @@ def test_get_target_data_df_error_cases():
 
 def test__max_as_of_le__reference_date():
     hub_dir = Path('tests/hubs/flu-metrocast')
-    hub_config = HubConfig(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
+    hub_config = HubConfigPtc(hub_dir, hub_dir / 'hub-config/predtimechart-config.yml')
     target_data_df = hub_config.get_target_data_df()
     for ref_date, exp_max_as_of in [('2025-01-29', None), ('2025-02-12', '2025-02-12'), ('2025-02-13', '2025-02-12'),
                                     ('2025-02-26', '2025-02-25')]:
